requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
DateTime==5.5
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
frozenlist==1.4.1
geoip2==4.8.0
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
maxminddb==2.6.2
multidict==6.0.5
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scapy==2.5.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
zope.interface==7.0.3
geocoder==1.38.1
python-dotenv==1.0.0
django-crontab==0.7.1